Finding 'm' nearest locations from 'n' users
The idea was to build a feature for the app of our company where it
would locate the nearest events with given user locations. This is a
complete API built by me which locates 'm' nearest neighbors to 'n' user
points and then locates the coordinates on a map. I used data from a
JSON file called 'events.json' which I can't disclose since it's
redirected to me from my company. The data file had data in a way where
individual coordinates of events were given with a unique ID. Cheers!